TestService remove

+++ src/main/java/pl/schodowski/CryptoPriceAlert/Crypto.java
@@ -24,6 +24,6 @@ public class Crypto {
     private long marketCap;
 
     @JsonProperty("total_volume")
-    private long totalVolume;
+    private float totalVolume;
 
 }
